fix tester custom image freeze, fix names not updating live in preset editor, add debug option to build script (#17)

+++ build.ps1
@@ -5,6 +5,8 @@ $dockerImage = "automask-build"
 
 $rids = @("win-x64", "linux-x64")
 
+$debug = $args -contains "--debug"
+
 $csprojVersion = ([xml](Get-Content "$PSScriptRoot/$project")).Project.PropertyGroup.Version
 $appVersion = "v$csprojVersion"
 
@@ -29,11 +31,16 @@ function Build($rid) {
             exit $LASTEXITCODE
         }
 
+        $dockerPublishArgs = "dotnet publish $project -c Release -r $rid --self-contained true /p:PublishAot=true -o /src/build/$rid"
+        if ($debug) {
+            $dockerPublishArgs += " /p:NativeDebugSymbols=true /p:StripSymbols=false"
+        }
+
         docker run --rm `
             -v "${PSScriptRoot}:/src" `
             -w /src `
             $dockerImage `
-            dotnet publish $project -c Release -r $rid --self-contained true /p:PublishAot=true -o /src/build/$rid `
+            sh -c $dockerPublishArgs `
             | Out-Host
 
         if ($LASTEXITCODE -ne 0) {
@@ -57,6 +64,11 @@ function Build($rid) {
         "-o", $output
     )
 
+    if ($debug) {
+        $dotnetArgs += "/p:NativeDebugSymbols=true"
+        $dotnetArgs += "/p:StripSymbols=false"
+    }
+
     dotnet @dotnetArgs | Out-Host
 
     if ($LASTEXITCODE -ne 0) {
@@ -64,8 +76,10 @@ function Build($rid) {
         exit $LASTEXITCODE
     }
 
-    # Remove debug symbols from output
-    Get-ChildItem "$output" -Filter "*.pdb" -ErrorAction SilentlyContinue | Remove-Item
+    # Remove debug symbols from output (unless --debug specified)
+    if (-not $debug) {
+        Get-ChildItem "$output" -Filter "*.pdb" -ErrorAction SilentlyContinue | Remove-Item
+    }
 
     Write-Host "  Output: $output"
     return $output
